requirements.txt
docker
requests
python-dotenv
kubernetes
pyyaml
gitpython
